svn:revision:39734
svn:branch:cmake
svn:account:starseeker

+++ src/libfb/fb_paged_io.c
@@ -34,20 +34,19 @@
 #include "fb.h"
 
 
-#define PAGE_BYTES	(63*1024L)		/* Max # of bytes/dma. */
-#define PAGE_PIXELS	(((PAGE_BYTES/sizeof(RGBpixel))/ifp->if_width) *ifp->if_width)
-#define PAGE_SCANS	(ifp->if_ppixels/ifp->if_width)
+#define PAGE_BYTES (63*1024L)		/* Max # of bytes/dma. */
+#define PAGE_PIXELS (((PAGE_BYTES/sizeof(RGBpixel))/ifp->if_width) *ifp->if_width)
+#define PAGE_SCANS (ifp->if_ppixels/ifp->if_width)
 
-#define Malloc_Bomb(_bytes_) \
-		fb_log("\"%s\"(%d) : allocation of %d bytes failed.\n", \
-				__FILE__, __LINE__, _bytes_)
-
+#define Malloc_Bomb(_bytes_)					\
+    fb_log("\"%s\"(%d) : allocation of %d bytes failed.\n",	\
+	   __FILE__, __LINE__, _bytes_)
 
 
 static int
 _fb_pgout(register FBIO *ifp)
 {
-    int scans, first_scan;
+    size_t scans, first_scan;
 
     /*fb_log("_fb_pgout(%d)\n", ifp->if_pno);*/
 
@@ -55,7 +54,7 @@ _fb_pgout(register FBIO *ifp)
 	return 1;
 
     first_scan = ifp->if_pno * PAGE_SCANS;
-    if (first_scan + PAGE_SCANS > ifp->if_height)
+    if (first_scan + PAGE_SCANS > (size_t)ifp->if_height)
 	scans = ifp->if_height - first_scan;
     else
 	scans = PAGE_SCANS;
@@ -68,10 +67,11 @@ _fb_pgout(register FBIO *ifp)
 	);
 }
 
+
 static int
 _fb_pgin(register FBIO *ifp, int pageno)
 {
-    int scans, first_scan;
+    size_t scans, first_scan;
 
     /*fb_log("_fb_pgin(%d)\n", pageno);*/
 
@@ -81,7 +81,7 @@ _fb_pgin(register FBIO *ifp, int pageno)
     ifp->if_pdirty = 0;
 
     first_scan = ifp->if_pno * PAGE_SCANS;
-    if (first_scan + PAGE_SCANS > ifp->if_height)
+    if (first_scan + PAGE_SCANS > (size_t)ifp->if_height)
 	scans = ifp->if_height - first_scan;
     else
 	scans = PAGE_SCANS;
@@ -94,6 +94,7 @@ _fb_pgin(register FBIO *ifp, int pageno)
 	);
 }
 
+
 static int
 _fb_pgflush(register FBIO *ifp)
 {
@@ -103,7 +104,7 @@ _fb_pgflush(register FBIO *ifp)
 
     if (ifp->if_pdirty) {
 	if (_fb_pgout(ifp) <= -1)
-	    return	-1;
+	    return -1;
 	ifp->if_pdirty = 0;
     }
 
@@ -134,13 +135,14 @@ fb_ioinit(register FBIO *ifp)
 	if ((ifp->if_pbase = (unsigned char *)malloc(ifp->if_ppixels * sizeof(RGBpixel)))
 	    == PIXEL_NULL) {
 	    Malloc_Bomb(ifp->if_ppixels * sizeof(RGBpixel));
-	    return	-1;
+	    return -1;
 	}
     }
     ifp->if_pcurp = ifp->if_pbase;	/* Initialize pointer.	*/
     ifp->if_pendp = ifp->if_pbase+ifp->if_ppixels*sizeof(RGBpixel);
     return 0;
 }
+
 
 int
 fb_seek(register FBIO *ifp, int x, int y)
@@ -155,16 +157,16 @@ fb_seek(register FBIO *ifp, int x, int y)
 
     if (x < 0 || y < 0 || x >= ifp->if_width || y >= ifp->if_height) {
 	fb_log("fb_seek: illegal address <%d, %d>.\n", x, y);
-	return	-1;
+	return -1;
     }
     pixelnum = ((long) y * (long) ifp->if_width) + x;
     pagepixel = (long) ifp->if_pno * ifp->if_ppixels;
     if (pixelnum < pagepixel || pixelnum >= (pagepixel + ifp->if_ppixels)) {
 	if (ifp->if_pdirty)
 	    if (_fb_pgout(ifp) == - 1)
-		return	-1;
+		return -1;
 	if (_fb_pgin(ifp, (int) (pixelnum / (long) ifp->if_ppixels)) <= -1)
-	    return	-1;
+	    return -1;
     }
     ifp->if_pixcur = pixelnum;
     /* Compute new pixel pointer into page. */
@@ -172,6 +174,7 @@ fb_seek(register FBIO *ifp, int x, int y)
 	(ifp->if_pixcur % ifp->if_ppixels)*sizeof(RGBpixel);
     return 0;
 }
+
 
 int
 fb_tell(register FBIO *ifp, int *xp, int *yp)
@@ -187,12 +190,13 @@ fb_tell(register FBIO *ifp, int *xp, int *yp)
     return 0;
 }
 
+
 int
 fb_wpixel(register FBIO *ifp, unsigned char *pixelp)
 {
     if (ifp->if_pno == -1)
 	if (_fb_pgin(ifp, ifp->if_pixcur / ifp->if_ppixels) <= -1)
-	    return	-1;
+	    return -1;
 
     COPYRGB((ifp->if_pcurp), pixelp);
     ifp->if_pcurp += sizeof(RGBpixel);	/* position in page */
@@ -200,30 +204,32 @@ fb_wpixel(register FBIO *ifp, unsigned char *pixelp)
     ifp->if_pdirty = 1;	/* page referenced (dirty) */
     if (ifp->if_pcurp >= ifp->if_pendp) {
 	if (_fb_pgout(ifp) <= -1)
-	    return	-1;
+	    return -1;
 	ifp->if_pno = -1;
     }
     return 0;
 }
+
 
 int
 fb_rpixel(register FBIO *ifp, unsigned char *pixelp)
 {
     if (ifp->if_pno == -1)
 	if (_fb_pgin(ifp, ifp->if_pixcur / ifp->if_ppixels) <= -1)
-	    return	-1;
+	    return -1;
 
     COPYRGB(pixelp, (ifp->if_pcurp));
     ifp->if_pcurp += sizeof(RGBpixel);	/* position in page */
     ifp->if_pixcur++;	/* position in framebuffer */
     if (ifp->if_pcurp >= ifp->if_pendp) {
 	if (_fb_pgout(ifp) <= -1)
-	    return	-1;
+	    return -1;
 	ifp->if_pno = -1;
     }
 
     return 0;
 }
+
 
 int
 fb_flush(register FBIO *ifp)
@@ -232,10 +238,11 @@ fb_flush(register FBIO *ifp)
 
     /* call device specific flush routine */
     if ((*ifp->if_flush)(ifp) <= -1)
-	return	-1;
+	return -1;
 
     return 0;
 }
+
 
 /*
  * Local Variables:
